pyproject.toml
[project]
name = "selfhosteddeploytest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "litestar-vite>=0.14.0",
    "litestar[sqlalchemy,standard]>=2.17.0",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
linting = [
    "pre-commit>=4.3.0",
    "ruff>=0.12.12",
]
